server/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-dateutil==2.8.2
cryptography==42.0.5 